main.go: fixes sending message with upload
Fixes issue: https://github.com/Labfox/whatsfly/issues/149

The problem with that issue is that InjectMessageWithUploadData tries to set fields on an uninitialized object (originMessage.ImageMessage for example), so the proposing fix creates ImageMessage into originMessage prior to setting their fields.

Another problem raised, is that SendMessageWithUploadWrapper tries to mimc SendMessageProtobufWrapper by unmarshaling c_message into waE2E.Message object, but if an upload is appended to it, the message is not delivered. The trick is that waE2E.Message must be empty and the c_message passes to be the caption of the upload data.